requirements.txt
# Core dependencies
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0

# MQTT
paho-mqtt==1.6.1

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Farmingpy (for soil analysis)
git+https://github.com/TwinYields/farmingpy.git@main

# Dashboard
streamlit==1.28.2
plotly==5.18.0
altair==5.1.2

# Utilities
PyYAML==6.0.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0

# Logging and monitoring
loguru==0.7.2
rich==13.5.2
